SWIM through the NATs
The goal of this project is to implement and evaluate a decentralized
membership service that provides information about all the the
participating
processes. A membership service can easily be implemented using
a heatbeat mechanism, where each process periodically sends a ping
message
to all the processes in the system. Failures are detected when a process
does not responds to the ping message. However, this solution does
not scale well when the number of processes in the system grow because
heatbeat based mechanisms requires O(n2) messages in each round.
The project is split into two phases. In the first phase of the project
you will implement SWIM: Scalable Weakly-consistent Infection-style
Group Membership Protocol [1] using Kompics that scales to at least
hundreds of peers in simulation. In phase two you will modify the system
to make with work on the Open Internet.